Safari-609.1.11.1/LayoutTests/svg/animations/use-animate-width-and-height-expected.txt

Test animation of use element width/height

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS use.width.animVal.value is 100
PASS use.width.baseVal.value is 100
PASS use.height.animVal.value is 100
PASS use.height.baseVal.value is 100
PASS use.getAttribute('width') is '100'
PASS use.getAttribute('height') is '100'
PASS useElementTargetClone().width.animVal.value is 100
PASS useElementTargetClone().height.animVal.value is 100
PASS use.width.animVal.value is 105
PASS use.width.baseVal.value is 100
PASS use.height.animVal.value is 105
PASS use.height.baseVal.value is 100
FAIL use.getAttribute('width') should be 105. Was 100.
FAIL use.getAttribute('height') should be 105. Was 100.
PASS useElementTargetClone().width.animVal.value is 105
PASS useElementTargetClone().height.animVal.value is 105
PASS use.width.animVal.value is 115
PASS use.width.baseVal.value is 100
PASS use.height.animVal.value is 115
PASS use.height.baseVal.value is 100
FAIL use.getAttribute('width') should be 115. Was 100.
FAIL use.getAttribute('height') should be 115. Was 100.
PASS useElementTargetClone().width.animVal.value is 115
PASS useElementTargetClone().height.animVal.value is 115
PASS use.width.animVal.value is 125
PASS use.width.baseVal.value is 100
PASS use.height.animVal.value is 125
PASS use.height.baseVal.value is 100
FAIL use.getAttribute('width') should be 125. Was 100.
FAIL use.getAttribute('height') should be 125. Was 100.
PASS useElementTargetClone().width.animVal.value is 125
PASS useElementTargetClone().height.animVal.value is 125
PASS use.width.animVal.value is 135
PASS use.width.baseVal.value is 100
PASS use.height.animVal.value is 135
PASS use.height.baseVal.value is 100
FAIL use.getAttribute('width') should be 135. Was 100.
FAIL use.getAttribute('height') should be 135. Was 100.
PASS useElementTargetClone().width.animVal.value is 135
PASS useElementTargetClone().height.animVal.value is 135
PASS successfullyParsed is true

TEST COMPLETE
